+++ src/fypp.py
@@ -2508,8 +2508,13 @@ class Fypp:
         if options.modules:
             self._import_modules(options.modules, evaluator, syspath,
                                  options.moduledirs)
+        evaluate = options.define_mode == 'eval'
         if options.defines:
-            self._apply_definitions(options.defines, evaluator)
+            self._apply_definitions(options.defines, evaluator, evaluate)
+        if options.defines_str:
+            self._apply_definitions(options.defines_str, evaluator, False)
+        if options.defines_eval:
+            self._apply_definitions(options.defines_eval, evaluator, True)
         if inspect.signature(parser_factory) == inspect.signature(Parser):
             parser = parser_factory(includedirs=options.includes,
                                     encoding=self._encoding)
@@ -2588,18 +2593,21 @@ class Fypp:
 
 
     @staticmethod
-    def _apply_definitions(defines, evaluator):
+    def _apply_definitions(defines, evaluator, evaluate):
         for define in defines:
             words = define.split('=', 1)
             name = words[0]
-            value = None
-            if len(words) > 1:
-                try:
-                    value = evaluator.evaluate(words[1])
-                except Exception as exc:
-                    msg = "exception at evaluating '{0}' in definition for " \
-                          "'{1}'".format(words[1], name)
-                    raise FyppFatalError(msg) from exc
+            if evaluate:
+                value = None
+                if len(words) > 1:
+                    try:
+                        value = evaluator.evaluate(words[1])
+                    except Exception as exc:
+                        msg = "exception at evaluating '{0}' in definition for " \
+                            "'{1}'".format(words[1], name)
+                        raise FyppFatalError(msg) from exc
+            else:
+                value = str(words[1]) if len(words) > 1 else ""
             evaluator.define(name, value)
 
 
@@ -2671,6 +2679,9 @@ class FyppOptions(optparse.Values):
     def __init__(self):
         optparse.Values.__init__(self)
         self.defines = []
+        self.define_mode = 'eval'
+        self.defines_str = []
+        self.defines_eval = []
         self.includes = []
         self.line_numbering = False
         self.line_numbering_mode = 'full'
@@ -2824,11 +2835,28 @@ def get_option_parser():
     parser = optparse.OptionParser(prog=fypp_name, description=fypp_desc,
                                    version=fypp_version, usage=usage)
 
-    msg = 'define variable, value is interpreted as ' \
-          'Python expression (e.g \'-DDEBUG=1\' sets DEBUG to the ' \
-          'integer 1) or set to None if omitted'
+    msg = 'define a variable; value interpreted depending on the '\
+          '--define-mode option, either evaluated as a Python expression (e.g '\
+          '\'-DDEBUG=1\' sets DEBUG to the integer 1) with None as default '\
+          'value, or as string literal with the empty string as default value'
     parser.add_option('-D', '--define', action='append', dest='defines',
                       metavar='VAR[=VALUE]', default=defs.defines, help=msg)
+
+    msg = 'value evaluation mode in -D options, \'eval\': values are evaluated as Python '\
+          'expressions (default, requires quotation for string values), '\
+          '\'str\': values are string literals (no quotation needed)'
+    parser.add_option('--define-mode', metavar='MODE', choices=['eval', 'str'],
+                      default=defs.define_mode, dest='define_mode', help=msg)
+
+    msg = 'define a variable with a string literal value; similar to the -D option, but the value '\
+          'is always interpreted as a string literal with the empty string as default value'
+    parser.add_option('-S', '--define-str', action='append', dest='defines_str',
+                      metavar='VAR[=VALUE]', default=defs.defines_str, help=msg)
+
+    msg = 'define a variable with an evaluated expression; similar to the -D option, but the '\
+          'value is always evaluated as a Python expression with None as default value'
+    parser.add_option('-E', '--define-eval', action='append', dest='defines_eval',
+                      metavar='VAR[=VALUE]', default=defs.defines_eval, help=msg)
 
     msg = 'add directory to the search paths for include files'
     parser.add_option('-I', '--include', action='append', dest='includes',