+++ CHANGELOG.rst
@@ -6,6 +6,22 @@ Change Log
 Unreleased
 ==========
 
+Added
+-----
+
+* ``--define-mode`` option to control how values in ``-D`` options are
+  interpreted: either as Python expressions (requiring proper quotation for
+  strings) or as string literals (no quotation needed).
+
+* ``-S`` / ``--define-str`` option to define variables similarly to ``-D``, but
+  with values always treated as string literals independent of the settings in
+  the ``--define-mode`` option.
+
+* ``-E`` / ``--define-eval`` option to define variables similarly to ``-D``, but
+  with values always evaluated as Python expressions independent of the settings
+  in the ``--define-mode`` option.
+
+
 Changed
 -------